big cleanup with new Doom bugfixes and modified main Fenster libraries

+++ examples/doom-c/p_maputl.c
@@ -846,7 +846,8 @@ static void InterceptsOverrun(int num_intercepts, intercept_t *intercept)
 
     InterceptsMemoryOverrun(location, intercept->frac);
     InterceptsMemoryOverrun(location + 4, intercept->isaline);
-    InterceptsMemoryOverrun(location + 8, (int) intercept->d.thing);
+    InterceptsMemoryOverrun(location + 8, (int)(uintptr_t)intercept->d.thing);
+//    InterceptsMemoryOverrun(location + 8, (int) intercept->d.thing);
 }
